+++ CustomRunOptions/docs/release-notes.md
@@ -4,6 +4,7 @@
 
 ## Upcoming release
 
+* Added options to force or block any boot sequence options.
 * Fixed the mod crashing on controllers.
 
 ## 1.0.0